(bugfix) Fixed a styling error
When the content container has additional style, for example a max-width, the topbar should have the same property. Before the commit, this was done by giving the App container an auto-width; but this caused a bug where other containers wouldn't take their full width. The solution is giving the top-bar an additional styling property that can optionally be used with the same styling properties as the content container.